chore: 미커밋 변경사항 일괄 반영 (docs, tests, scripts, reports)
- AGENTS.md: GitHub Actions 파이프라인 명세 정정, Python 테스트 규칙 추가
- claude.md: 제목·배포 승인 규칙 업데이트
- test_rules.md: Python 테스트 실행 순서 추가
- unit_tests/: 테스트 구조 재편 (01_pairUSDT, 02_backend 분리)
- scripts/verify_021_incremental.sh: 021 증분 업데이트 검증 스크립트 추가
- reports/49.md, reports/51.md: 작업 보고서 추가

Closes #61
