+++ reports/49.md
@@ -1,0 +1,26 @@
+# #49 Render 배포 환경 설정 개선 보고서
+
+## 개요
+Render 배포를 위한 백엔드 환경 설정 개선
+
+## 변경 파일
+
+### 1. `02_backend/requirements.txt` — 의존성 버전 고정
+- `>=` 방식 → `==` 방식으로 변경해 빌드 재현성 확보
+- `fastapi==0.115.6`
+- `uvicorn[standard]==0.34.0`
+- `supabase==2.11.0`
+- `python-dotenv==1.0.1`
+- `starlette==0.41.3` (신규 추가)
+- `websockets==15.0.1` (신규 추가)
+
+### 2. `02_backend/.python-version` — Python 버전 명시 (신규)
+- Render가 기본값으로 빌드하는 문제 방지
+- `3.13` 고정
+
+### 3. `render.yaml` — `CYCLE_COIN_ID` 환경변수 추가
+- `routers/btc_chart.py`에서 사용 중이나 누락되어 있던 항목 등록
+
+## 결과
+- PR #50 생성 및 머지 진행
+- Closes #49
