+++ reports/51.md
@@ -1,0 +1,22 @@
+# #51 frontend Vercel 배포 환경 설정 및 API URL 정리 보고서
+
+## 개요
+03_frontend, 04_frontend_0222fb Vercel 배포를 위한 코드 정리
+
+## 변경 파일
+
+### 03_frontend/src/LegacyAnalyzerApp.tsx
+- sample-dashboard-data.json 폴백 제거
+- API 실패 시 에러 throw로 단순화
+
+### 04_frontend_0222fb/src/lib/api.js
+- 하드코딩된 구 Render URL(`btcreack20260222-fb.onrender.com`) 제거
+- `VITE_API_URL` 환경변수 단일 경로로 통일
+
+## Vercel 배포 설정 (대시보드에서 직접 설정 필요)
+- 03_frontend → Root Directory: `03_frontend`, Env: `VITE_API_URL=https://two0260317-alt.onrender.com`
+- 04_frontend_0222fb → Root Directory: `04_frontend_0222fb`, Env: `VITE_API_URL=https://two0260317-alt.onrender.com`
+
+## 결과
+- PR 생성 및 main 머지
+- Closes #51
