+++ unit_tests/01_pairUSDT/lib/predictor/test_predict.py
