+++ unit_tests/01_pairUSDT/lib/predictor/test_predict_cycle_box_count.py
